pyproject.toml
[tool.poetry]
name = "langpify"
version = "0.1.0"
description = "SDK para programación basada en agentes neuro-simbolicos"
authors = ["Patricio Gerpe <pj.patriciojulian@gmail.com>"]
readme = "README.md"
repository = "https://github.com/codewithpatelo/langpify"
documentation = "https://langpify.readthedocs.io"
homepage = "https://github.com/codewithpatelo/langpify"
keywords = ["ai", "agents", "llm", "neuro-symbolic", "pii", "privacy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
packages = [{include = "langpify", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
presidio-analyzer = "^2.2.0"
presidio-anonymizer = "^2.2.0"
scrubadub = "^2.0.0"
typing-extensions = {version = "^4.0.0", python = "<3.10"}
langgraph = "^0.0.60"
python-a2a = "^0.1.0"

[tool.poetry.scripts]
build = "poetry.core.masonry.api:build"
publish = "poetry.core.masonry.api:publish"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.3.0"
pytest-cov = "^4.1.0"
python-dotenv = "^1.0.0"
openai = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
